fix input edit edu periode

+++ views/pages/page_varOption/index.php
@@ -54,7 +54,11 @@
                             <td><?=$dt->catatan?></td>
                             <td><?=$dt->parent?></td>
                             <td><a class="btn btn-info" href="#" onclick="edit(<?= $dt->recid; ?>)" data-bs-toggle="modal" data-bs-target="#editModal"><i class="fas fa-info-circle"></i></a>
+                           <?php if($dt->disabled == true){ ?>
+                            <a class="btn btn-danger disabled"><i class="fas fa-trash"></i></a>
+                            <?php } else { ?>
                             <a class="btn btn-danger" href="/hewi/public/var/delete/<?= $dt->recid; ?>" onclick="return confirm('yakin ingin hapus data?')"><i class="fas fa-trash"></i></a>
+                            <?php } ?>
                             </td>
                         </tr>
                         <?php endforeach ?>
